plugin-builder: init
Add a library function and a ruby script for automatically generating
plugin attrsets.

A "plugins" file will be a YAML file like this:

	github:
		- owner/repo
		- …
	some-other-provider:
		- …

This file is processed by `prefetch-plugins.rb` to generate an
importable lockfile for all of the plugins. This value can be passed to
buildPlugins to make the plugin attrset.